UserJpaDaoTest done

+++ paw-webapp/persistence/src/main/java/ar/edu/itba/paw/persistence/UserJpaDao.java
@@ -1,12 +1,10 @@
 package ar.edu.itba.paw.persistence;
 
-import ar.edu.itba.paw.model.Audition;
 import ar.edu.itba.paw.model.FilterOptions;
 import ar.edu.itba.paw.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
-
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
@@ -150,7 +148,7 @@ public class UserJpaDao implements UserDao {
 
     private List<Long> getUserIds(Query query) {
         @SuppressWarnings("unchecked")
-        List<Long> ids = (List<Long>) query.getResultList().stream().map(o -> ((Integer) o).longValue()).collect(Collectors.toList());
+        List<Long> ids = (List<Long>) query.getResultList().stream().map(o -> ((BigInteger) o).longValue()).collect(Collectors.toList());
 
         if(ids.isEmpty())
             ids.add(-1L);
